moved AddUsosAuth  to extension
